fix(test): avoid DeprecationWarning

+++ src/__tests__/modules/extensionManager.test.ts
@@ -24,7 +24,7 @@ beforeAll(async () => {
 
 afterEach(() => {
   disposeAll(disposables)
-  if (fs.existsSync(tmpfolder)) {
+  if (tmpfolder && fs.existsSync(tmpfolder)) {
     fs.rmSync(tmpfolder, { force: true, recursive: true })
   }
 })
